refactor(producer_fault): split panic-hook silencing into a child module
Pure move of PanicHook/silence_injected_panics/SilencedInjectedPanics into
producer_fault/silence.rs, re-exported so every caller keeps its existing
producer_fault::silence_injected_panics path. Keeps producer_fault.rs under the
~800-line campsite target now that the #3154 construction seam lives there too.
